DBZ-7024 Move RecordProcessors into separate classes
