+++ debezium-embedded/src/main/java/io/debezium/embedded/async/AsyncEmbeddedEngine.java
@@ -53,7 +53,6 @@ import org.slf4j.LoggerFactory;
 
 import io.debezium.DebeziumException;
 import io.debezium.annotation.VisibleForTesting;
-import io.debezium.common.annotation.Incubating;
 import io.debezium.config.CommonConnectorConfig;
 import io.debezium.config.Configuration;
 import io.debezium.config.Instantiator;
@@ -971,55 +970,6 @@ public final class AsyncEmbeddedEngine<R> implements DebeziumEngine<R>, AsyncEng
     }
 
     /**
-     * Generalization of {@link DebeziumEngine.ChangeConsumer}, giving the user complete control over the whole records processing chain.
-     * Processor is initialized with all the required engine internals, like chain of transformations, to be able to implement whole record processing chain.
-     * Implementations can provide serial or parallel processing of the change records or anything in between, eventually also add any other kind of manipulation
-     * with the records.
-     * Any exception thrown during the processing of the records it propagated to the caller.
-     */
-    @Incubating
-    public interface RecordProcessor<R> {
-
-        /**
-         * Initialize the processor with objects created and managed by {@link DebeziumEngine}, which are needed for records processing.
-         *
-         * @param recordService {@link ExecutorService} which allows to run processing of individual records in parallel.
-         * @param transformations chain of transformations to be applied on every individual record.
-         * @param committer implementation of {@link DebeziumEngine.RecordCommitter} responsible for committing individual records as well as batches.
-         */
-        void initialize(ExecutorService recordService, Transformations transformations, RecordCommitter committer);
-
-        /**
-         * Processes a batch of records provided by the source connector.
-         * Implementations are assumed to use {@link DebeziumEngine.RecordCommitter} to appropriately commit individual records and the batch itself.
-         *
-         * @param records List of {@link SourceRecord} provided by the source connector to be processed.
-         * @throws Exception Any exception is propagated to the caller.
-         */
-        void processRecords(List<SourceRecord> records) throws Exception;
-    }
-
-    /**
-     * Abstract implementation of {@link RecordProcessor}, which provides implementation of processor initialization, while the record processing implementation
-     * left to the children classes.
-     */
-    private abstract class AbstractRecordProcessor implements RecordProcessor<R> {
-        protected ExecutorService recordService;
-        protected Transformations transformations;
-        protected RecordCommitter committer;
-
-        @Override
-        public void initialize(final ExecutorService recordService, final Transformations transformations, final RecordCommitter committer) {
-            this.recordService = recordService;
-            this.transformations = transformations;
-            this.committer = committer;
-        }
-
-        @Override
-        public abstract void processRecords(List<SourceRecord> records) throws Exception;
-    }
-
-    /**
      * Default completion callback which just logs the error. If connector finishes successfully it does nothing.
      */
     private static class DefaultCompletionCallback implements DebeziumEngine.CompletionCallback {
@@ -1070,293 +1020,6 @@ public final class AsyncEmbeddedEngine<R> implements DebeziumEngine<R>, AsyncEng
             final Configuration config = Configuration.from(task.context().config());
             return DelayStrategy.exponential(Duration.ofMillis(config.getInteger(EmbeddedEngineConfig.ERRORS_RETRY_DELAY_INITIAL_MS)),
                     Duration.ofMillis(config.getInteger(EmbeddedEngineConfig.ERRORS_RETRY_DELAY_MAX_MS)));
-        }
-    }
-
-    /**
-     * {@link Callable} which applies transformations to a single record.
-     */
-    private static class TransformRecord implements Callable<SourceRecord> {
-        private final SourceRecord record;
-        private final Transformations transformations;
-
-        TransformRecord(final SourceRecord record, final Transformations transformations) {
-            this.record = record;
-            this.transformations = transformations;
-        }
-
-        @Override
-        public SourceRecord call() {
-            final SourceRecord transformedRecord = transformations.transform(record);
-            return transformedRecord != null ? transformedRecord : null;
-        }
-    }
-
-    /**
-     * {@link Callable} which applies transformations to a single record and convert the record into desired format.
-     */
-    private class TransformAndConvertRecord<R> implements Callable<R> {
-        private final SourceRecord record;
-        private final Transformations transformations;
-        private final Function<SourceRecord, R> converter;
-
-        TransformAndConvertRecord(final SourceRecord record, final Transformations transformations, final Function<SourceRecord, R> converter) {
-            this.record = record;
-            this.transformations = transformations;
-            this.converter = converter;
-        }
-
-        @Override
-        public R call() {
-            final SourceRecord transformedRecord = transformations.transform(record);
-            return transformedRecord != null ? converter.apply(transformedRecord) : null;
-        }
-    }
-
-    /**
-     * {@link Callable} which applies transformations to a single record and thereafter passes it to provided consumer.
-     */
-    private static class TransformAndConsumeRecord implements Callable<Void> {
-        private final SourceRecord record;
-        private final Transformations transformations;
-        private final Consumer<SourceRecord> consumer;
-
-        TransformAndConsumeRecord(final SourceRecord record, final Transformations transformations, final Consumer<SourceRecord> consumer) {
-            this.record = record;
-            this.transformations = transformations;
-            this.consumer = consumer;
-        }
-
-        @Override
-        public Void call() {
-            final SourceRecord transformedRecord = transformations.transform(record);
-            if (transformedRecord != null) {
-                consumer.accept(transformedRecord);
-            }
-            return null;
-        }
-    }
-
-    /**
-     * {@link Callable} which applies transformations to a single record, transformed it into desired format and applies provided use consumer on this record.
-     */
-    private static class TransformConvertConsumeRecord<R> implements Callable<Void> {
-        private final SourceRecord record;
-        private final Transformations transformations;
-        private final Function<SourceRecord, R> serializer;
-        private final Consumer<R> consumer;
-
-        TransformConvertConsumeRecord(final SourceRecord record, final Transformations transformations, final Function<SourceRecord, R> serializer,
-                                      final Consumer<R> consumer) {
-            this.record = record;
-            this.transformations = transformations;
-            this.serializer = serializer;
-            this.consumer = consumer;
-        }
-
-        @Override
-        public Void call() {
-            final SourceRecord transformedRecord = transformations.transform(record);
-            if (transformedRecord != null) {
-                consumer.accept(serializer.apply(transformedRecord));
-            }
-            return null;
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which runs transformations of the records in parallel and then pass the whole batch to the user-provided handler.
-     * This processor should be used when user provides its own {@link DebeziumEngine.ChangeConsumer} and records shouldn't be converted to different format.
-     */
-    private class ParallelSmtBatchProcessor extends AbstractRecordProcessor {
-        final DebeziumEngine.ChangeConsumer<SourceRecord> userHandler;
-
-        ParallelSmtBatchProcessor(final DebeziumEngine.ChangeConsumer<SourceRecord> userHandler) {
-            this.userHandler = userHandler;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<SourceRecord>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformRecord(r, transformations))));
-
-            LOGGER.trace("Thread {} is getting source records.", Thread.currentThread().getName());
-            final List<SourceRecord> transformedRecords = new ArrayList<>(recordFutures.size());
-            for (Future<SourceRecord> f : recordFutures) {
-                SourceRecord record = f.get(); // we need the whole batch, eventually wait forever
-                if (record != null) {
-                    transformedRecords.add(record);
-                }
-            }
-
-            LOGGER.trace("Calling user handler.");
-            userHandler.handleBatch(transformedRecords, committer);
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which transforms and converts the records in parallel and then pass the whole batch to the user-provided handler.
-     * This processor should be used when user provides its own {@link DebeziumEngine.ChangeConsumer} and records should be converted to different format.
-     */
-    private class ParallelSmtAndConvertBatchProcessor extends AbstractRecordProcessor {
-        final DebeziumEngine.ChangeConsumer<R> userHandler;
-        final Function<SourceRecord, R> convertor;
-
-        ParallelSmtAndConvertBatchProcessor(final DebeziumEngine.ChangeConsumer<R> userHandler, final Function<SourceRecord, R> convertor) {
-            this.userHandler = userHandler;
-            this.convertor = convertor;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<R>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformAndConvertRecord<R>(r, transformations, convertor))));
-
-            LOGGER.trace("Getting source records.");
-            final List<R> convertedRecords = new ArrayList<>(recordFutures.size());
-            for (Future<R> f : recordFutures) {
-                R record = f.get(); // we need the whole batch, eventually wait forever
-                if (record != null) {
-                    convertedRecords.add(record);
-                }
-            }
-
-            LOGGER.trace("Calling user handler.");
-            userHandler.handleBatch(convertedRecords, committer);
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which transforms the records in parallel. Records are passed to the user-provided {@link Consumer}.
-     * This processor should be used when user provides only custom {@link Consumer} and records should be passed without converting to the consumer in the same
-     * order as they were obtained from the database.
-     */
-    private class ParallelSmtConsumerProcessor extends AbstractRecordProcessor {
-        final Consumer<SourceRecord> consumer;
-
-        ParallelSmtConsumerProcessor(final Consumer<SourceRecord> consumer) {
-            this.consumer = consumer;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<SourceRecord>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformRecord(r, transformations))));
-
-            LOGGER.trace("Waiting for the batch to finish processing.");
-            final List<SourceRecord> transformedRecords = new ArrayList<>(recordFutures.size());
-            for (Future<SourceRecord> f : recordFutures) {
-                transformedRecords.add(f.get()); // we need the whole batch, eventually wait forever
-            }
-
-            LOGGER.trace("Calling user consumer.");
-            for (int i = 0; i < records.size(); i++) {
-                consumer.accept(transformedRecords.get(i));
-                committer.markProcessed(records.get(i));
-            }
-
-            LOGGER.trace("Marking batch as finished.");
-            committer.markBatchFinished();
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which transforms and converts the records in parallel. Converted records are passed to the user-provided {@link Consumer}.
-     * This processor should be used when user provides only custom {@link Consumer}, records should be converted and passed to the consumer in the same order as they
-     * were obtained from the database.
-     */
-    private class ParallelSmtAndConvertConsumerProcessor extends AbstractRecordProcessor {
-        final Consumer<R> consumer;
-        final Function<SourceRecord, R> convertor;
-
-        ParallelSmtAndConvertConsumerProcessor(final Consumer<R> consumer, final Function<SourceRecord, R> convertor) {
-            this.consumer = consumer;
-            this.convertor = convertor;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<R>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformAndConvertRecord(r, transformations, convertor))));
-
-            LOGGER.trace("Waiting for the batch to finish processing.");
-            final List<R> convertedRecords = new ArrayList<>(recordFutures.size());
-            for (Future<R> f : recordFutures) {
-                convertedRecords.add(f.get()); // we need the whole batch, eventually wait forever
-            }
-
-            LOGGER.trace("Calling user consumer.");
-            for (int i = 0; i < records.size(); i++) {
-                consumer.accept(convertedRecords.get(i));
-                committer.markProcessed(records.get(i));
-            }
-
-            LOGGER.trace("Marking batch as finished.");
-            committer.markBatchFinished();
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which transforms the records in parallel. Records are passed to the user-provided {@link Consumer} in arbitrary order, once they are
-     * processed. This processor should be used when user provides only custom {@link Consumer} and records should be passed without converting to the consumer in the same
-     * order as they were obtained from the database.
-     */
-    private class ParallelSmtAsyncConsumerProcessor extends AbstractRecordProcessor {
-        final Consumer<SourceRecord> consumer;
-
-        ParallelSmtAsyncConsumerProcessor(final Consumer<SourceRecord> consumer) {
-            this.consumer = consumer;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<Void>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformAndConsumeRecord(r, transformations, consumer))));
-
-            LOGGER.trace("Waiting for the batch to finish processing.");
-            for (int i = 0; i < records.size(); i++) {
-                recordFutures.get(i);
-                committer.markProcessed(records.get(i));
-            }
-
-            LOGGER.trace("Marking batch as finished.");
-            committer.markBatchFinished();
-        }
-    }
-
-    /**
-     * {@link RecordProcessor} which transforms and converts the records in parallel. Records are passed to the user-provided {@link Consumer} in arbitrary order, once
-     * they are processed. This processor should be used when user provides only custom {@link Consumer}, records should be converted and passed to the consumer in
-     * arbitrary order.
-     */
-    private class ParallelSmtAndConvertAsyncConsumerProcessor extends AbstractRecordProcessor {
-        final Consumer<R> consumer;
-        final Function<SourceRecord, R> convertor;
-
-        ParallelSmtAndConvertAsyncConsumerProcessor(final Consumer<R> consumer, final Function<SourceRecord, R> convertor) {
-            this.consumer = consumer;
-            this.convertor = convertor;
-        }
-
-        @Override
-        public void processRecords(final List<SourceRecord> records) throws Exception {
-            LOGGER.debug("Thread {} is submitting {} records for processing.", Thread.currentThread().getName(), records.size());
-            final List<Future<Void>> recordFutures = new ArrayList<>(records.size());
-            records.stream().forEachOrdered(r -> recordFutures.add(recordService.submit(new TransformConvertConsumeRecord<>(r, transformations, convertor, consumer))));
-
-            LOGGER.trace("Waiting for the batch to finish processing.");
-            for (int i = 0; i < records.size(); i++) {
-                recordFutures.get(i);
-                committer.markProcessed(records.get(i));
-            }
-
-            LOGGER.trace("Marking batch as finished.");
-            committer.markBatchFinished();
         }
     }
